+++ common/src/main/java/fr/loudo/narrativecraft/client/narrative/character/ClientNpcEditor.java
@@ -78,6 +78,7 @@ public class ClientNpcEditor implements ClientNarrativeEntryEditor<NpcPayload, N
             } catch (Exception ignored) {
             }
         }
+        npc.setCustomNbt(payload.getCustomNbt());
 
         UtilsClient.reloadListScreen();
     }
@@ -115,6 +116,7 @@ public class ClientNpcEditor implements ClientNarrativeEntryEditor<NpcPayload, N
             npc.setModelType(Utils.parsePlayerModelType(payload.getModelType()));
         }
         npc.setEntityType(resolveEntityType(payload.getEntityTypeId()));
+        npc.setCustomNbt(payload.getDialogDataJson());
         String dialogDataJson = payload.getDialogDataJson();
         if (dialogDataJson != null && !dialogDataJson.isEmpty() && !dialogDataJson.equals("{}")) {
             try {