+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Timed choice
 - Image tag
 - Target a character using @char(name) in the command tag (e.g. `# command "attribute @char(user) minecraft:scale base set 3"`)
+- Assign custom nbt to characters
 
 ### Fixes
 - Dialog 2D scale not fixed based on gui scale